Updated to load all data.
git-svn-id: https://xp-dev.com/svn/MapleGame@42 22f4eea3-df18-4490-a9d7-79e6856d3c20

+++ ChannelServer/Maple/Life/LifeObject.cs
@@ -16,8 +16,8 @@ namespace Loki.Maple.Life
         public LifeObject(dynamic datum)
             : base()
         {
-            this.ID = datum.id;
-            this.MapleID = datum.lifeid;
+            this.ID = (int)datum.id;
+            this.MapleID = (int)datum.lifeid;
 
             this.Position = new Point(datum.x_pos, datum.y_pos);
             this.Foothold = datum.foothold;
